refactor: convert put.js to TypeScript
Convert put.js to TypeScript with proper type definitions, removing the
need for eslint-disable workarounds. Define interfaces for app parameters
(PutApp, PathApp, NotificationApp) that specify only the properties each
function actually uses.

Use branded types (Context, Path, SourceRef) from @signalk/server-api
for type-safe delta construction, following existing codebase patterns.

Add missing 'self' property to FullSignalK type declaration to match
runtime behavior.
